Odds Api Swagger - Fix OddsApiService and related data types
