+++ odds-api-swagger/src/main/java/betapi/oddsapiservice/model/Event.java
@@ -3,7 +3,6 @@ package betapi.oddsapiservice.model;
 import betapi.oddsapiservice.utils.Utils;
 
 import java.util.Date;
-import java.util.List;
 
 public class Event {
 
@@ -13,7 +12,6 @@ public class Event {
     private String commence_time;
     private String home_team;
     private String away_team;
-    private List<Bookmaker> bookmakers;
 
     public String getId() {
         return id;
@@ -67,13 +65,5 @@ public class Event {
 
     public void setAway_team(String away_team) {
         this.away_team = away_team;
-    }
-
-    public List<Bookmaker> getBookmakers() {
-        return bookmakers;
-    }
-
-    public void setBookmakers(List<Bookmaker> bookmakers) {
-        this.bookmakers = bookmakers;
     }
 }
