+++ odds-api-swagger/src/main/java/betapi/oddsapiservice/model/EventOdds.java
@@ -1,0 +1,16 @@
+package betapi.oddsapiservice.model;
+
+import java.util.List;
+
+public class EventOdds extends Event{
+
+    private List<Bookmaker> bookmakers;
+
+    public List<Bookmaker> getBookmakers() {
+        return bookmakers;
+    }
+
+    public void setBookmakers(List<Bookmaker> bookmakers) {
+        this.bookmakers = bookmakers;
+    }
+}
